Add client and server middleware
Here's the basic functionality.

Using the client middleware allows any job to be set to a given corral
queue using Sidekiq::Job.set or by setting Sidekiq::Corral.current
before enqueueing the job.

Then the server middleware sets Sidekiq::Corral.current based on the
corral for the job being processed, and that's used to put any jobs that
are enqueued while the initial job is being processed into the same
queue referenced in the corral.