Add Corral.confine
There will likely be places where you'll want to use a corral for a
block of of code either because the jobs being enqueued are happening
further down a call stack or because the top level job is a descendant
of ActiveJob instead of Sidekiq::Worker (which can use .set(corral:)).
This convenience method will help encourage us to scope corral usage to
a specific bit of code instead of having to remember to remove the
current value when we're done.

I considered making Corral.current= a private class method to really
drive that point home, but I'm going to default to public for now.