Add exempt_queues option
While Sidekiq::Corral is meant to confine all jobs and child/side-effect
jobs into a single queue, some infrastructure rules may include special
queues that we don't want to override. This may be a queue that talks to
a particular external system and we might want to pause if that system
is having issues. Or it could include something where the latency
requirements are extremely important.

This exempt_queues option (yes, I thought about continuing with the
corral-related language, but it started to feel like it was getting out
of hand!) lets you pass an array of these special queues to
Sidekiq::Corral.install or any time you add the Client middleware.

And when we go to enqueue a job, Corral will skip overriding the job
queue if it's set to one of these exempt queues.

It will still pass the corral with the job, so any further jobs enqueued
by the exempt queue job will still be eligible to be put on the corral
queue. That is, a job using an exempt queue doesn't break the corral
chain entirely.