Move tests to use actual Sidekiq jobs
Until now, all of the tests were making some (founded) assumptions about
how Sidekiq works and uses middleware. In order to build more confidence
in the whole setup, especially how corrals are carried through sub-jobs,
we can actually add the middleware in test and enqueue and run jobs, the
process of which exercises the middleware stack (as long as we've set it
up).

Because the Sidekiq fake testing setup doesn't use a separate server
process, we have to add our server middleware to a separate testing
middleware chain. That works well, but it means that we aren't
particularly testing the client middleware being added to the server
side. To me, that's a fine trade-off as it's such a well-trodden path.

The other big benefit this gives us is that the tests more clearly serve
as documentation, showing how you'd actually use Corral in your own
code, because it sets the corral using Sidekiq's own APIs (e.g.
SomeJob.set(corral: "critical")).