Make Corral.current= private
Setting Corral.current has some very wide-ranging effects, enqueuing all
jobs on a specific queue, so the stakes for forgetting to set it back to
`nil` (or any other existing value!) when you're done are much higher.
We'd definitely want to avoid hard to track down bugs where we think we
reset it properly but didn't, and now another bit of work in the same
thread enqueues jobs in a queue we don't want them in.

Taking Corral.current= private makes it practically more difficult to
get into those situations. It also communicates that the way you should
be setting this value directly-if at all-is via Corral.confine, which
has the key bits to reset the value properly.