Use options hash for Corral middleware
Sidekiq middleware doesn't support using keyword arguments, but
positional arguments don't give much context to what we're configuring
in practice. And since the one option Corral takes is an exception, it's
particularly confusing.

Let's bite the bullet and use an options hash, which can get messy, but
the clarity is a big benefit. It also leaves a cleaner opening for other
options down the road (e.g. exempt job classes).